+++ google-scripts/.claspignore
@@ -1,1 +1,2 @@
 __tests__/*
+__mocks__/*
